backend/requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
mysql-connector-python==8.1.0
pymongo==4.4.0
pytest==7.4.0
pytest-flask==1.2.0
